LIBASPACE-369. Updated "Department Codes" styling for resources
The "resource_identifiers.js" file is broadly similar to the
"accession_identifiers.js" file (which itself was derived from the
"yale_accession_identifiers.js" file.

In v1.1, the HTML styling for the "Department Codes" drop-down was
updated in the "yale_accession_identifiers.js", so incorporated the
update into the "resource_identifiers.js" (it has already been
incorporated into the "accession_identifiers.js" file.

https://umd-dit.atlassian.net/browse/LIBASPACE-369